lower MAX-RETRIES; work around open problem
Scale MAX-RETRIES back to something that Rakudo could reasonably do before
the programmer throws up his hands in exasperation waiting for tempfile()
to return if there's a problem.

open() seems to die when it can't open the file.  Wrap it in a try block
so that we can capture its death.